Allow custom subtype to ElementName
